feat(exceptions): adding some validations, and exceptions, also fixing a possivble use of input to map who is user and who is not

+++ src/main/java/com/example/schoolservlet/daos/SchoolClassDAO.java
@@ -1,6 +1,10 @@
 package com.example.schoolservlet.daos;
 
 import com.example.schoolservlet.daos.interfaces.GenericDAO;
+import com.example.schoolservlet.exceptions.DataException;
+import com.example.schoolservlet.exceptions.InvalidNumberException;
+import com.example.schoolservlet.exceptions.NotFoundException;
+import com.example.schoolservlet.exceptions.RequiredFieldException;
 import com.example.schoolservlet.models.SchoolClass;
 import com.example.schoolservlet.utils.PostgreConnection;
 
@@ -11,22 +15,22 @@ import java.util.Map;
 public class SchoolClassDAO implements GenericDAO<SchoolClass> {
 
     @Override
-    public boolean create(SchoolClass schoolClass) {
+    public void create(SchoolClass schoolClass) throws DataException, RequiredFieldException {
+        if (schoolClass.getSchoolYear() == null || schoolClass.getSchoolYear().isEmpty()) throw new RequiredFieldException("nome da turma");
         try(Connection conn = PostgreConnection.getConnection();
             PreparedStatement pstmt = conn.prepareStatement("INSERT INTO school_class (school_year) VALUES (?)")){
 
             pstmt.setString(1, schoolClass.getSchoolYear());
 
-            return pstmt.executeUpdate() > 0;
-
+            pstmt.executeUpdate();
         } catch(SQLException sqle){
             sqle.printStackTrace();
-            return false;
+            throw new DataException("Erro ao criar turma", sqle);
         }
     }
 
     @Override
-    public Map<Integer, SchoolClass> findMany(int skip, int take) {
+    public Map<Integer, SchoolClass> findMany(int skip, int take) throws DataException {
         Map<Integer, SchoolClass> schoolClassMap = new HashMap<>();
 
         try (Connection conn = PostgreConnection.getConnection();
@@ -45,13 +49,15 @@ public class SchoolClassDAO implements GenericDAO<SchoolClass> {
 
         } catch (SQLException sqle){
             sqle.printStackTrace();
+            throw new DataException("Erro ao listar turmas", sqle);
         }
 
         return schoolClassMap;
     }
 
     @Override
-    public SchoolClass findById(int id){
+    public SchoolClass findById(int id) throws InvalidNumberException, DataException, NotFoundException {
+        if (id <= 0) throw new InvalidNumberException("id", "ID deve ser maior do que 0");
         try(
                 Connection conn = PostgreConnection.getConnection();
                 PreparedStatement pstmt = conn.prepareStatement("SELECT id, school_year FROM school_class WHERE id = ?")
@@ -65,45 +71,48 @@ public class SchoolClassDAO implements GenericDAO<SchoolClass> {
                         rs.getInt("id"),
                         rs.getString("school_year")
                 );
-            }
-
+            } else throw new NotFoundException("turma", "id", String.valueOf(id));
         } catch (SQLException sqle) {
             sqle.printStackTrace();
+            throw new DataException("Erro ao buscar turma");
         }
-
-        return null;
     }
 
     @Override
-    public boolean update(SchoolClass schoolClass) {
+    public void update(SchoolClass schoolClass) throws DataException, InvalidNumberException, NotFoundException, RequiredFieldException{
+        if (schoolClass.getId() <= 0) throw new InvalidNumberException("id", "ID deve ser maior do que 0");
+        if (schoolClass.getSchoolYear() == null || schoolClass.getSchoolYear().isEmpty()) throw new RequiredFieldException("nome da turma");
+
         try (Connection conn = PostgreConnection.getConnection();
              PreparedStatement pstmt = conn.prepareStatement("UPDATE school_class SET school_year = ? " +
                      "WHERE id = ?")){
             pstmt.setString(1, schoolClass.getSchoolYear());
             pstmt.setInt(2, schoolClass.getId());
 
-            return pstmt.executeUpdate() > 0;
+            if (pstmt.executeUpdate() <= 0) throw new NotFoundException("turma", "id", String.valueOf(schoolClass.getId()));
         } catch (SQLException sqle){
             sqle.printStackTrace();
-            return false;
+            throw new DataException("Erro ao atualizar turma");
         }
     }
 
     @Override
-    public boolean delete(int id) {
+    public void delete(int id) throws InvalidNumberException, DataException, NotFoundException {
+        if (id <= 0) throw new InvalidNumberException("id", "ID deve ser maior do que 0");
+
         try(Connection conn = PostgreConnection.getConnection();
             PreparedStatement pstmt = conn.prepareStatement("DELETE FROM school_class WHERE id = ?")){
             pstmt.setInt(1, id);
 
-            return pstmt.executeUpdate() > 0;
+            if (pstmt.executeUpdate() <= 0) throw new NotFoundException("turma", "id", String.valueOf(id));
         } catch (SQLException sqle){
             sqle.printStackTrace();
-            return false;
+            throw new DataException("Erro ao deletar turma", sqle);
         }
     }
 
     @Override
-    public int totalCount(){
+    public int totalCount() throws DataException {
         try(Connection conn = PostgreConnection.getConnection();
             PreparedStatement pstmt = conn.prepareStatement("SELECT COUNT(*) AS totalCount FROM school_class");
             ResultSet rs = pstmt.executeQuery()){
@@ -113,6 +122,7 @@ public class SchoolClassDAO implements GenericDAO<SchoolClass> {
             }
         } catch (SQLException sqle){
             sqle.printStackTrace();
+            throw new DataException("Erro ao contar turmas", sqle);
         }
 
         return -1;
